feat containers/array: move array to containers/static directory
