+++ main.c
@@ -4,7 +4,7 @@
 #include <containers/dynamic/string/string.h>
 #include <containers/dynamic/string/string_impl.h>
 
-#include <tests/containers/array/array.h>
+#include <tests/containers/static/array/array.h>
 #include <tests/containers/dynamic/string/string.h>
 #include <tests/utils/result/result.h>
 #include <tests/fs/file/encoding.h>
