+++ src/containers/static/CMakeLists.txt
@@ -3,5 +3,4 @@ target_sources(${PROJECT_NAME}
         common.h
 )
 
-add_subdirectory(dynamic)
-add_subdirectory(static)
+add_subdirectory(array)
