+++ src/containers/static/array/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(${PROJECT_NAME}
+    PRIVATE
+        array.h
+)
