+++ src/containers/static/common.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <containers/common.h>
+
+#define NAMESPACE_CONTAINERS_STATIC(NAME) NAMESPACE_CONTAINERS(CONCAT3(static, __, NAME))
