+++ src/tests/containers/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_subdirectory(array)
+add_subdirectory(static)
 add_subdirectory(dynamic)
