+++ src/tests/containers/static/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(array)
